Reorganise core
