+++ SmartSolar.Device.Core/Common/GpioOutputConnection.cs
@@ -1,23 +1,21 @@
 ﻿using Windows.Devices.Gpio;
 using Caliburn.Micro;
-using SmartSolar.Device.Core.Domain;
 
-namespace SmartSolar.Device.Core.Services
+namespace SmartSolar.Device.Core.Common
 {
 	/// <summary>
-	/// Single responsibility: represent a physical output connection to something, via a GPIO pin
+	///     Single responsibility: represent a physical output connection to something, via a GPIO pin
 	/// </summary>
 	public class GpioOutputConnection : PropertyChangedBase, IOutputConnection
 	{
 		private GpioPin _gpioPin;
-		private bool? _state = null;
+		private bool? _state;
 
 		public bool? State
 		{
-			get { return _state; } 
+			get { return _state; }
 			set
 			{
-				
 				if (value == _state) return;
 				_state = value;
 				_gpioPin.Write(GpioPinValue.High);
