--- SmartSolar.Device.Core/Domain/GpioPumpOutputConnection.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SmartSolar.Device.Core.Services;
-
-namespace SmartSolar.Device.Core.Domain
-{
-	public class GpioPumpOutputConnection: GpioOutputConnection , IPumpOutputConnection 
-	{
-	}
-}
