--- SmartSolar.Device.Core/Domain/IPumpOutputConnection.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SmartSolar.Device.Core.Domain
-{
-	public interface IPumpOutputConnection: IOutputConnection
-	{
-	}
-}
