+++ SmartSolar.Device.Core/Pump/FakePumpOutputConnection.cs
@@ -1,0 +1,8 @@
+﻿using SmartSolar.Device.Core.Common;
+
+namespace SmartSolar.Device.Core.Pump
+{
+	public class FakePumpOutputConnection: FakeOutputConnection, IPumpOutputConnection
+	{
+	}
+}
