+++ SmartSolar.Device.Core/Pump/IPumpOutputConnection.cs
@@ -1,0 +1,8 @@
+﻿using SmartSolar.Device.Core.Common;
+
+namespace SmartSolar.Device.Core.Pump
+{
+	public interface IPumpOutputConnection: IOutputConnection
+	{
+	}
+}
