+++ SmartSolar.Device.Core/Sensor/ThermistorTemperatureReader.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SmartSolar.Device.Core.Domain;
+using SmartSolar.Device.Core.Common;
 
-namespace SmartSolar.Device.Core.Services
+namespace SmartSolar.Device.Core.Sensor
 {
 	/// <summary>
 	/// Single responsibility: read the current temperature of a thermistor connected via an ADC.
