--- SmartSolar.Device.Core/Services/FakePumpOutputConnection.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SmartSolar.Device.Core.Domain;
-
-namespace SmartSolar.Device.Core.Services
-{
-	public class FakePumpOutputConnection: FakeOutputConnection, IPumpOutputConnection
-	{
-	}
-}
